refactor: Move DTO to oauth's dto package from order's dto package
