+++ src/main/java/gift/oauth/KakaoOauthService.java
@@ -6,8 +6,10 @@ import static org.springframework.http.MediaType.APPLICATION_FORM_URLENCODED;
 import gift.exception.FailedLoginException;
 import gift.exception.InvalidAccessTokenException;
 import gift.member.MemberService;
-import gift.order.dto.DefaultMessageTemplate;
-import gift.order.dto.KakaoUserInfoDTO;
+import gift.oauth.config.KakaoOauthConfigure;
+import gift.oauth.message.DefaultMessageTemplate;
+import gift.oauth.dto.KakaoTokenResponseDTO;
+import gift.oauth.dto.KakaoUserInfoDTO;
 import java.net.URI;
 import java.util.Map;
 import java.util.Objects;
