+++ src/main/java/gift/oauth/config/KakaoOauthConfigure.java
@@ -1,4 +1,4 @@
-package gift.oauth;
+package gift.oauth.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
