+++ src/main/java/gift/oauth/dto/KakaoTokenResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.oauth;
+package gift.oauth.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
