+++ src/main/java/gift/oauth/dto/KakaoUserInfoDTO.java
@@ -1,4 +1,4 @@
-package gift.order.dto;
+package gift.oauth.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
